ch-#: modify frontend component -update profile page ui -update header component -update viewall page
