+++ src/components/DetailsAction/hooks/useChatClick.ts
@@ -1,4 +1,4 @@
-import { useState } from "react"
+import { useCallback, useState } from "react"
 
 const useChatClick = ()=>{
     const [isChatClick, setIsChatClick] = useState<boolean>(false)
@@ -7,7 +7,9 @@ const useChatClick = ()=>{
         setIsChatClick(value)
     }
 
-    return {isChatClick, handleChatClick}
+    const handleChatClickCalback = useCallback(handleChatClick, [handleChatClick])
+
+    return {isChatClick, handleChatClick,handleChatClickCalback}
 }
 
 export default useChatClick
