+++ src/components/Filter/hooks/useCloseFilter.ts
@@ -1,7 +1,11 @@
 import { useRef } from "react"
+import { useSearchParams } from "react-router-dom"
 
 const useCloseFilter = (setIsFilterClicked: React.Dispatch<React.SetStateAction<boolean>>,setFilter: React.Dispatch<React.SetStateAction<string>>)=>{
     const ref = useRef<HTMLDivElement | null>(null)
+
+
+    const [searchParams, setSearchParams]= useSearchParams()
 
     const handleSlideDown = ()=>{
         if(ref && ref.current){
@@ -13,8 +17,9 @@ const useCloseFilter = (setIsFilterClicked: React.Dispatch<React.SetStateAction<
     const handleFilter = (filterType: string)=>{
         setFilter(filterType)
         setIsFilterClicked(false)
+        setSearchParams({label: filterType})
     }
 
-    return {handleFilter,handleSlideDown, ref}
+    return {handleFilter,handleSlideDown, ref, searchParams}
 }
 export default useCloseFilter
