+++ src/components/Filterslider/type.ts
@@ -1,0 +1,7 @@
+type TFilter = {
+    setFilter: React.Dispatch<React.SetStateAction<string>>,
+    setIsFilterClicked: React.Dispatch<React.SetStateAction<boolean>>
+}
+
+
+export type {TFilter}
