+++ src/components/Header/index.tsx
@@ -25,7 +25,7 @@ const Header = ({children, title}: THeader)=>{
                 <Link to="/profile"><li>Profile</li></Link>
              <li onClick={()=> logOut()}>logout</li> 
             </div>
-            <div className="icon-left" onClick={()=> navigate(-1)}>{children[0]}</div>
+            <div className="icon-left" onClick={()=> navigate("/home")}>{children[0]}</div>
             <div className="page-title">{title}</div>
             {children[2] ? children[2] : ""}
             <div className="icon-right" onClick={(e)=> onOpenNav(e)} > {children[1]} </div>
